+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- \[brew,cask\] Disable analytics and env hints in logs.
+
 ## {gh}`4.10.0 (2022-03-31) <compare/v4.9.10...v4.10.0>`
 
 - \[mpm\] Allow for package managers to simultaneously set a list of